src/Parser/requirements.txt
selenium>=4.0.0
beautifulsoup4>=4.10.0
requests>=2.28.0
psycopg2-binary>=2.9.3
